During completion we show symbols from the same system first.
This is extremely useful when you are working on package-inferred system
with large number of packages. For example, working on package ultralisp/widgets/landing
I want to complete get-name function which is in the ultralisp/models/user package.
Before this commit, autocomplete showed me the relevant result ultralisp/models/user:get-name
on a 37 position in the list. After this commit, it will be shown on first position,
because the symbol is in the same primary system as I'm working on.

This change caused a performace degradation on a first call, because now sly has to
cache a mapping package -> primary system for every known package.